+++ react/src/index.ts
@@ -27,4 +27,5 @@ export * from './useBaseURL';
 export * from './useCurrentAgent';
 export * from './useDebounce';
 export * from './useLocalStorage';
+export * from './useMarkdown';
 export * from './useServerSearch';